[BEAM-8433] DataCatalogBigQueryIT runs for both Calcite and ZetaSQL dialects.
